requirements.txt
gspread==5.12.0
oauth2client==4.1.3
pandas>=2.2.0
python-dotenv==1.0.0
pyyaml==6.0.1
schedule==1.2.0
requests==2.31.0
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.36
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
Pillow==10.1.0

